// examples/next-js-sdk-gen-2-experimental-app-directory/src/sdk/blocks/image/image.types.ts
import type { BuilderBlock } from '../../types/builder-block';
export interface ImageProps {
  attributes?: string;
  highPriority?: boolean;
  className?: string;
  image: string;
  sizes?: string;
  lazy?: boolean;
  height?: number;
  width?: number;
  altText?: string;
  backgroundSize?: 'cover' | 'contain';
  backgroundPosition?: string;
  srcset?: string;
  aspectRatio?: number;
  lockAspectRatio?: boolean;
  children?: any;
  fitContent?: boolean;
  builderBlock?: BuilderBlock;
  noWebp?: boolean;
  src?: string;
}